+++ cea/demand/metamodel/nn_generator/nn_random_sampler.py
@@ -1,10 +1,6 @@
 # coding=utf-8
 """
-'nn_pipeline.py" script is a pipeline of the following jobs:
-    (1) calls "sampling_main" function for random generation of features
-    (2) calls "neural_trainer" function for training a first neural network and saving the model
-    (3) executes a loop in which "sampling_main" and "neural_training" are iteratively called for
-        sequential training of the neural network.
+'nn_random_sampler.py' script is a generator of random properties for the entire case-study
 """
 
 __author__ = "Fazel Khayatian"
@@ -22,17 +18,41 @@ from cea.demand.calibration.bayesian_calibrator.calibration_sampling import appl
 from cea.demand import demand_main
 import pickle
 import cea
-import json
 #import h5py
 import os
 import numpy as np
 import pandas as pd
-from cea.demand.calibration.nn_generator.nn_settings import number_samples_scaler, random_variables,\
+from cea.demand.metamodel.nn_generator import number_samples, random_variables,\
     target_parameters, boolean_vars
-from cea.demand.calibration.nn_generator.input_prepare import input_prepare_main
+from cea.demand.metamodel.nn_generator import input_prepare_main
+# import h5py
+import os
+import pickle
+
+import numpy as np
+import pandas as pd
+
+import cea
+from cea.demand import demand_main
+from cea.demand.calibration.bayesian_calibrator.calibration_sampling import apply_sample_parameters
+from cea.demand.calibration.latin_sampler import latin_sampler
+from cea.demand.metamodel.nn_generator import input_prepare_main
+from cea.demand.metamodel.nn_generator import number_samples, random_variables, \
+    target_parameters, boolean_vars
+from cea.demand.demand_main import properties_and_schedule
 
 
-def sampling_scaler(locator, random_variables, target_parameters, list_building_names, weather_path, gv):
+def input_dropout(urban_input_matrix, urban_taget_matrix):
+    rows, cols = urban_input_matrix.shape
+    drop_random_array=np.random.rand(rows)
+    drop_idx_filter=drop_random_array>0.5
+    drop_idx=np.where(drop_idx_filter)
+    urban_input_matrix=np.delete(urban_input_matrix,drop_idx,0)
+    urban_taget_matrix=np.delete(urban_taget_matrix,drop_idx,0)
+
+    return urban_input_matrix, urban_taget_matrix
+
+def sampling_main(locator, random_variables, target_parameters, list_building_names, weather_path, gv):
     '''
     this function creates a number of random samples for the entire district (city)
     :param locator: points to the variables
@@ -48,9 +68,9 @@ def sampling_scaler(locator, random_variables, target_parameters, list_building_
     size_city = np.shape(list_building_names)
     size_city=size_city[0]
     #   create random samples of the entire district
-    for i in range(number_samples_scaler): #the parameter "number_samples" is accessible from 'nn_settings.py'
+    for i in range(number_samples): #the parameter "number_samples" is accessible from 'nn_settings.py'
         bld_counter=0
-        # create list of samples with a LHC sampler and save to disk
+        # create list of samples with a LHC sampler and save to disk (*.csv)
         samples, pdf_list = latin_sampler(locator, size_city, random_variables)
         for building_name in (list_building_names):
             np.save(locator.get_calibration_samples(building_name), samples)
@@ -76,23 +96,23 @@ def sampling_scaler(locator, random_variables, target_parameters, list_building_
             overwritten.loc[overwritten.Name == building_name, random_variables] = sample[:,1]
             bld_counter = bld_counter + 1
 
-        # for boolean_mask in (boolean_vars):
-        #     fazel1=overwritten.replace([0],'False')
-        #     overwritten[boolean_mask] = fazel1
-        #     overwritten[boolean_mask] = overwritten[boolean_mask].replace(1, 'True')
+        #   write to csv format
         overwritten.to_csv(locator.get_building_overrides())
 
-        # run cea demand
-
+        #   run cea demand
         demand_main.demand_calculation(locator, weather_path, gv)
+        #   prepare the inputs for feeding into the neural network
         urban_input_matrix, urban_taget_matrix=input_prepare_main(list_building_names, locator, target_parameters, gv)
-
-        scaler_inout_path = locator.get_minmaxscaler_folder()
-        file_path_inputs=os.path.join(scaler_inout_path,"input%(i)s.csv" % locals())
+        #   drop half the inputs and targets to avoid overfitting and save RAM / Disk space
+        urban_input_matrix, urban_taget_matrix=input_dropout(urban_input_matrix, urban_taget_matrix)
+        #   get the pathfor saving the files
+        nn_inout_path = locator.get_nn_inout_folder()
+        #   save inputs with sequential naming
+        file_path_inputs=os.path.join(nn_inout_path,"input%(i)s.csv" % locals())
         data_file_inputs = pd.DataFrame(urban_input_matrix)
         data_file_inputs.to_csv(file_path_inputs,header=False,index=False)
-
-        file_path_targets = os.path.join(scaler_inout_path, "target%(i)s.csv" % locals())
+        #   save inputs with sequential naming
+        file_path_targets = os.path.join(nn_inout_path, "target%(i)s.csv" % locals())
         data_file_targets = pd.DataFrame(urban_taget_matrix)
         data_file_targets.to_csv(file_path_targets,header=False,index=False)
 
@@ -106,7 +126,7 @@ def run_as_script():
     weather_path = locator.get_default_weather()
     building_properties, schedules_dict, date = properties_and_schedule(gv, locator)
     list_building_names = building_properties.list_building_names()
-    sampling_scaler(locator, random_variables, target_parameters, list_building_names, weather_path, gv)
+    sampling_main(locator, random_variables, target_parameters, list_building_names, weather_path, gv)
