+++ cea/demand/metamodel/nn_generator/nn_trainer.py
@@ -14,7 +14,7 @@ __status__ = "Production"
 
 from sklearn.preprocessing import MinMaxScaler
 from sklearn.externals import joblib
-from cea.demand.calibration.nn_generator.nn_settings import target_parameters
+from cea.demand.metamodel.nn_generator import target_parameters
 import numpy as np
 import pandas as pd
 from keras.layers import Input, Dense
@@ -25,9 +25,9 @@ import os
 from keras.models import Sequential
 from keras.callbacks import EarlyStopping
 import cea
-from cea.demand.calibration.nn_generator.input_prepare import input_prepare_main
+from cea.demand.metamodel.nn_generator import input_prepare_main
 from cea.demand.demand_main import properties_and_schedule
-from cea.demand.calibration.nn_generator.nn_settings import number_samples
+from cea.demand.metamodel.nn_generator import number_samples
 import theano
 import multiprocessing
